backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-ratelimit>=4.1,<5.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
whitenoise>=6.6,<7.0
gunicorn>=21.0,<22.0
ruff>=0.1.0
pytest-django>=4.5,<5.0
